requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
notion-client>=2.2.1
python-dotenv>=1.0.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
apscheduler>=3.10.0
yt-dlp>=2024.1.0
reportlab>=4.0.0
